refactor: 컴포넌트는 무조건 index에서 가져오도록 수정
